fix(governance): create usePDA and move function related info and pass Buffer library
